authentication added

+++ controllers/feed.js
@@ -4,6 +4,7 @@ const path = require('path');
 const { validationResult } = require('express-validator');
 
 const Post = require('../models/post');
+const User = require('../models/user');
 
 const clearImage = filePath => {
     filePath = path.join(__dirname, '..', filePath);
@@ -55,13 +56,23 @@ exports.createPost = (req, res, next) => {
         title,
         content,
         imageUrl,
-        creator: { name: 'Igor' },
+        creator: req.userIds,
     });
+    let creator;
     post.save()
+        .then(result => {
+            return User.findById(req.userId);
+        })
+        .then(user => {
+            creator = user;
+            user.posts.push(post);
+            return user.save();
+        })
         .then(result => {
             res.status(201).json({
                 message: 'Post created successfully!',
-                post: result
+                post,
+                creator: { _id: creator._id, name: creator.name }
             });
         })
         .catch(err => {
@@ -117,6 +128,11 @@ exports.updatePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error;
             }
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error('Not authorized!');
+                error.statusCode = 403;
+                throw error;
+            }
             if (imageUrl !== post.imageUrl) {
                 clearImage(post.imageUrl);
             }
@@ -145,8 +161,20 @@ exports.deletePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error;
             }
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error('Not authorized!');
+                error.statusCode = 403;
+                throw error;
+            }
             clearImage(post.imageUrl);
             return Post.findByIdAndRemove(postId);
+        })
+        .then(result => {
+            return User.findById(req.userId);
+        })
+        .then(user => {
+            user.posts.pull(postId);
+            return user.save();
         })
         .then(result => {
             res.status(200).json({ message: 'Post deleted!' });
